Install network-runner so that it includes all the network providers
I ran into the issue on the new cluster because the neutron image did not
include the dellemc os9 provider. Manually copying the os9 driver didn't fix
the issue because network_runners' old version of inventory.py does not
dynamically load the network providers (so in the logs it showed up as
"dellemcos9.os9 is an invalid choice").

Closes: https://github.com/CCI-MOC/esi/issues/490